The CSS file has been moved to the page folder. Breadcrumb names have been corrected.
